requirements.txt
# Core scientific computing
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
scipy>=1.11

# Visualization
matplotlib>=3.7
seaborn>=0.12

# HTTP clients (LLM API calls — no heavy SDK required)
requests>=2.31

# Secrets
python-dotenv>=1.0

# Output validation
pydantic>=2.0
jsonschema>=4.21

# Progress
tqdm>=4.65

# Jupyter
jupyter>=1.0
notebook>=7.0
ipykernel>=6.0

# YAML (model matrix config)
pyyaml>=6.0

# Optional: OpenAI SDK (providers use raw requests by default)
# openai>=1.0

# Optional: FastText baseline (Rule 05A only — large, ~7 GB unpacked)
# fasttext-wheel>=0.9.2
# gensim>=4.3
